C implementation: all instructions done.
All instructions have now been implemented (including full I/O support) but most of them are completely untested at this stage. Performance is good. It remains to be seen whether or not the NaN scheme turns out to be pleasant and convenient. If not, will roll back to trapping instead.