Decision: FVM 2.0 shall be big-endian.
This is the opposite to FVM 1.0. The reasons for the decision are: (1) To conform to standard network byte order (RFC 1700); (2) To be more easily understood by humans; (3) So that data stored in data memory, in the hold, in rom and in external files piped to or from the FVM will all appear in the same format and thus be directly comparable by humans (and indeed programmatically) without difficulty. Henceforth, data memory, the hold, rom, and stream I/O shall all be big-endian. The actual endianness of stack elements is irrelevant provided that this holds true. Thus it is simple and easy for computations to be done by a little-endian CPU, as indeed is demonstrated in the current use of 'fvm2.c' on x86 in development.